pyproject.toml
[project]
name = "console-voice-assistant"
version = "0.1.0"
description = "Asistente de voz por consola con Deepgram STT/TTS + OpenAI + barge-in"
requires-python = ">=3.11"
dependencies = [
    "deepgram-sdk==3.11.0",
    "openai>=1.40.0",
    "sounddevice>=0.4.6",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.2.1",
    "elevenlabs>=2.32.0",
    "torch>=1.12.0",
    "torchaudio>=0.12.0",
    "langchain>=1.2.7",
    "langgraph>=1.0.7",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[tool.hatch.build.targets.wheel]
packages = ["src/voice_pipeline", "src/agents", "src/api"]
sources = ["src"]

[dependency-groups]
dev = [
    "grandalf>=0.8",
    "ipykernel>=7.1.0",
    "langgraph-cli[inmem]>=0.4.12",
]



